add CRUD functions of work detail page.

+++ app/Http/Controllers/WorkController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Work;
 use App\Tag;
 use App\WorkTag;
+use App\WorkDetail;
 use Illuminate\Support\Facades\Storage;
 use Log;
 
@@ -42,9 +43,10 @@ class WorkController extends Controller
         return view('admin.works', ['works'=>$works, "tags"=>$tags, 'tagCur'=>$tag]);
     }
 
-	public function detail()
-    {
-        return view('wk_details');
+	public function detail($workId=null)
+	{
+		$details = WorkDetail::where('work_id',$workId)->get();
+		return view('wk_details', ['details'=>$details]);
     }
 
 	public function edit(Work $work)
@@ -57,52 +59,64 @@ class WorkController extends Controller
 	public function store(Request $request)
 	{
 		Log::info($request);
-		$this->validate($request, [
+		$validate = [
 			'fileField' => 'image|max:150|mimes:jpeg,png,gif,svg|dimensions:max_width=640,max_height=430',
 			'name_cn' => 'required|max:15',
 			'name_en' => 'required|max:40',
 			'time' => 'required',
-		]);
+		];
+		if(!$request->id){
+			$validate['fileField'] .= '|required';
+		}
+		$this->validate($request, $validate);
 
 		$work = Work::find($request->id);
 		if(!$work){
 			$work=new Work();
-			//$work->save();
-			Log::info($work);
 		}
 
 		$newpath=null;
 
 		if($request->hasFile('fileField')){
 			$oldpath = $work->thumb;
-			Log::info('old:'.$oldpath);
 
 			$fileName = md5($request->fileField).'.'.$request->file('fileField')->extension();
-			var_dump($request->fileField);
 			$path = $request->file('fileField')->storeAs('public/works'/*.$work->id*/, $fileName);
-			Log::info('path:'.$path);
 			$newpath = '/storage'.substr($path, 6);
 
 			if($oldpath && $oldpath != $newpath)
 				Storage::delete('/public'.substr($work->thumb, 8));
 		}
-
+		
 		$work->thumb=$newpath?$newpath:$work->thumb;
 		$work->name_cn=$request->name_cn;
 		$work->name_en=$request->name_en;
 		$work->time=$request->time;
 		$work->save();
 
-		WorkTag::where('work_id',$work->id)->delete();
-		foreach($request->tags as $tag=>$value){
-			$wt = new WorkTag();
-			$wt->work_id = $work->id;
-			$wt->tag_id = $tag;
-			$wt->save();
+		if($request->hasFile('fileFields')){
+			foreach($request->fileFields as $file){
+				$fileName = md5($file).'.'.$file->extension();
+				$path = $file->storeAs('public/works'/*.$work->id*/, $fileName);
+				$path = '/storage'.substr($path, 6);
+				$wd = new WorkDetail();
+				$wd->work_id = $work->id;
+				$wd->path = $path;
+				$wd->save();
+			}
 		}
-Log::info($work);
 
-        return redirect('/admin/works');
+		if($request->tags){
+			WorkTag::where('work_id',$work->id)->delete();
+			foreach($request->tags as $tag=>$value){
+				$wt = new WorkTag();
+				$wt->work_id = $work->id;
+				$wt->tag_id = $tag;
+				$wt->save();
+			}
+		}
+
+        return $request->id?back()->withInput():redirect('/admin/works_edit/'.$work->id);
 	}
 
 	public function del(Request $request)
@@ -110,11 +124,25 @@ Log::info($work);
 		$work = Work::find($request->id);
 		$this->delFile($work->thumb);
 		$work->delete();
+		
+		$workDeatils = WorkDetail::where('work_id',$request->id)->get();
+		foreach($workDeatils as $workDeatil){
+			$this->delFile($work->path);
+			$workDeatil->delete();
+		}
         return redirect('/admin/works');
+	}
+
+	public function delPic(Request $request)
+	{
+		$detail = WorkDetail::find($request->id);
+		$this->delFile($detail->path);
+		$detail->delete();
+        return response()->json(['ret'=>true]);
 	}
 
 	private function delFile($path)
 	{
-		Storage::delete('/public'.substr($path, 8));
+		return Storage::delete('/public'.substr($path, 8));
 	}
 }
